refactor: split drag drop helpers
